Refactor models, update migration
